WineSettings: set WINEPREFIX to `<compatdata>/pfx` when enabling a compat tool
This is here to work around legendary's broken handling for `STEAM_COMPAT_DATA_PATH`
when resolving save paths. The `LegendaryCore.get_save_path` method checks
if `STEAM_COMPAT_DATA_PATH` is set only on a per-game basis and it doesn't
fall back to checking the if the variable is set globally.